adding light documentation
