+++ app/src/main/java/com/example/clicker/presentation/stream/StreamView.kt
@@ -191,7 +191,7 @@ fun StreamView(
                                 )
                             }else{
                                 StreamManagerUI(
-                                    showEditStreamInfo={
+                                    showModView={
                                         streamViewModel.setAutoModSettings(false)
                                         showStreamManager()
                                     },
