+++ app/src/main/java/com/example/clicker/presentation/stream/views/streamManager/StreamManager.kt
@@ -30,11 +30,12 @@ import com.example.clicker.R
 
 
 /**
- * StreamManagerUI shows the user 3 buttons to press and display the wanted information
+ * StreamManagerUI is what is shown to the users that are in vertical portrait mode and click the mod icon to the right of the
+ * chat messages
  * */
 @Composable
 fun StreamManagerUI(
-    showEditStreamInfo:()->Unit,
+    showModView:()->Unit,
 
 ){
     Column(
@@ -43,12 +44,12 @@ fun StreamManagerUI(
 
         EditStreamInfoCard(
             title="Mod View ",
-            contentDescription = "Edit AutoMod Info",
+            iconContentDescription = "Edit AutoMod Info",
             iconPainter = painterResource(R.drawable.mod_view_24),
             iconTextColor =Color.White,
             showStreamManager={
                 //this conditional has been moved to now mean show modview
-                showEditStreamInfo()
+                showModView()
             }
         )
     }
@@ -56,10 +57,22 @@ fun StreamManagerUI(
 
 }
 
+
+/**
+ * EditStreamInfoCard is meant to show a individual option avaliable to the user when presented with the [StreamManagerUI] composable
+ *
+ * @param title Represents the main piece of information shown to the user. Should be no more than three words(preferably two)
+ * and tells the user the information they are getting when they clicking
+ * @param iconContentDescription represents the description of the [Icon] to be displayed next to [title]
+ * @param iconPainter represents the visual aspect of the [Icon] to be displayed next to [title]
+ * @param iconTextColor represents the color of the [Icon] to be displayed next to [title]
+ * @param showStreamManager represents the action to be taken when the surrounding card gets clicked.
+ *
+ * */
 @Composable
 fun EditStreamInfoCard(
     title:String,
-    contentDescription:String,
+    iconContentDescription:String,
     iconPainter: Painter,
     iconTextColor:Color,
     showStreamManager:()->Unit,
@@ -85,12 +98,11 @@ fun EditStreamInfoCard(
 
             Icon(
                 painter = iconPainter,
-                contentDescription,
+                contentDescription=iconContentDescription,
                 tint= iconTextColor,
                 modifier = Modifier.size(35.dp)
             )
             Text(text =title, color = iconTextColor, fontSize = MaterialTheme.typography.headlineMedium.fontSize)
-
         }
     }
 }
